pyproject.toml
[project]
name = "ai-psy-benchmark"
version = "1.0.0"
description = "AI Psychology Benchmark for evaluating AI models"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.8",
    "numpy>=2.4.2",
    "openai>=2.21.0",
    "pandas>=3.0.0",
    "plotly>=6.5.2",
    "polars>=1.38.1",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "rich>=14.3.2",
    "scipy>=1.17.0",
]
